+++ HISTORY.md
@@ -6,6 +6,7 @@
 * new dev version
 * Blocks: fixed a scope issue introduced in v4.2.2.4
 * Blocks: enabled "duplicate block definition" in the palette
+* updated German translation
 
 ## v4.2.2.5
 ###2018-11-02